issue37 playlist: also replay last clip in list with default settings
The special treatment of the last clip in a flowplayer playlist
must be worked around to achieve consistent behavior with the playlist
plugin's default behavior of replaying the current clip when clicking
Replay.

stop() in the onBeforeFinish event would jump to the first frame, so
we call pause() and add an onResume event which checks the current
position: if it's below 1 second before end, resuming will trigger
a play(lastclip).