issue39: make file extensions configurable
For both validExtensions and posterExtensions offer the option to
configure with a regex as string.

At the same time make the resulting regexes stricter:
- no 'g' flag
- anchor at end and beginning
- require a dot as first char

For the latter again bear in mind that the ipad plugin has a different
idea of extension than the Flash clip.extension property (no dot at
beginning). Does not come into play here, as Flash clip.extension is
read-only, or rather setting it explicitly has no effect.